rag-backend/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
pydantic>=2.0
python-multipart>=0.0.6
lightrag-hku>=0.0.9
httpx>=0.27
tiktoken>=0.7
numpy>=1.26
